+++ CHANGELOG.md
@@ -10,6 +10,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Add Template Element and Template Parts
 * Make all properties available as scss variables during the build
 * Update sass-loader to get rid of warnings
+* Add features to design object
+* Add security obj to design object
+* Add withRawObject function to design object
 
 
 ## [2.0.0] - 09.10.2025